refactor: rename files in build, map-adapter, shape-api, styler-store (#1199)
- build: progressAdapters.ts → progressAdapterUtils.ts (P2: functions-only)
- map-adapter: ports.ts → portTypes.ts (P2: types-only)
- shape-api: defaults.ts → defaultConstants.ts (P2: constants-only)
- styler-store: StylerAPI.ts → stylerApiTypes.ts (P2: types-only)

Rollback: git revert <hash> --no-edit
