+++ packages/map-adapter/src/index.ts
@@ -1,4 +1,4 @@
 export * from './types.js';
-export * from './ports.js';
+export * from './portTypes.js';
 export * from './MapViewService.js';
 export * from './adapters/MapLibreDeckAdapter.js';
